+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - **Download results error**. [#465](https://github.com/crawlab-team/crawlab/issues/465)
 - **crawlab-sdk CLI error**. [#458](https://github.com/crawlab-team/crawlab/issues/458)
 - **Page refresh issue**. [#441](https://github.com/crawlab-team/crawlab/issues/441)
+- **Results not support JSON**. [#202](https://github.com/crawlab-team/crawlab/issues/202)
 - **Getting all spider after deleting a spider**.
 - **i18n warning**.